+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/FeatureGatesTest.java
@@ -21,16 +21,8 @@ public class FeatureGatesTest {
     @ParallelTest
     public void testIndividualFeatureGates() {
         for (FeatureGates.FeatureGate gate : FeatureGates.NONE.allFeatureGates()) {
-            FeatureGates enabled;
-            FeatureGates disabled;
-
-            if ("UseKRaft".equals(gate.getName()))  {
-                enabled = new FeatureGates("+" + gate.getName() + ",+KafkaNodePools");
-                disabled = new FeatureGates("-" + gate.getName() + ",+KafkaNodePools");
-            } else {
-                enabled = new FeatureGates("+" + gate.getName());
-                disabled = new FeatureGates("-" + gate.getName());
-            }
+            FeatureGates enabled = new FeatureGates("+" + gate.getName());
+            FeatureGates disabled = new FeatureGates("-" + gate.getName());
 
             assertThat(enabled.allFeatureGates().stream().filter(g -> gate.getName().equals(g.getName())).findFirst().orElseThrow().isEnabled(), is(true));
             assertThat(disabled.allFeatureGates().stream().filter(g -> gate.getName().equals(g.getName())).findFirst().orElseThrow().isEnabled(), is(false));
@@ -60,14 +52,14 @@ public class FeatureGatesTest {
 
     @ParallelTest
     public void testFeatureGatesParsing() {
-        assertThat(new FeatureGates("+UseKRaft,+KafkaNodePools").useKRaftEnabled(), is(true));
+        assertThat(new FeatureGates("+UseKRaft").useKRaftEnabled(), is(true));
         assertThat(new FeatureGates("-StableConnectIdentities").stableConnectIdentitiesEnabled(), is(false));
         assertThat(new FeatureGates("+KafkaNodePools").kafkaNodePoolsEnabled(), is(true));
         assertThat(new FeatureGates("-UseKRaft,-StableConnectIdentities").useKRaftEnabled(), is(false));
         assertThat(new FeatureGates("-UseKRaft,-StableConnectIdentities").stableConnectIdentitiesEnabled(), is(false));
         assertThat(new FeatureGates("-UseKRaft,-StableConnectIdentities,-KafkaNodePools").kafkaNodePoolsEnabled(), is(false));
         assertThat(new FeatureGates("  +UseKRaft    ,    +KafkaNodePools").useKRaftEnabled(), is(true));
-        assertThat(new FeatureGates("  +UseKRaft    ,    +KafkaNodePools").kafkaNodePoolsEnabled(), is(true));
+        assertThat(new FeatureGates("  -UseKRaft    ,    -KafkaNodePools").kafkaNodePoolsEnabled(), is(false));
         assertThat(new FeatureGates("+StableConnectIdentities,-UseKRaft").useKRaftEnabled(), is(false));
         assertThat(new FeatureGates("+StableConnectIdentities,-UseKRaft").stableConnectIdentitiesEnabled(), is(true));
     }
@@ -114,15 +106,13 @@ public class FeatureGatesTest {
 
     @ParallelTest
     public void testKraftAndKafkaNOdePoolsNotFulfilled() {
-        InvalidConfigurationException e = assertThrows(InvalidConfigurationException.class, () -> new FeatureGates("+UseKRaft"));
-        assertThat(e.getMessage(), containsString("The UseKRaft feature gate can be enabled only together with the KafkaNodePools feature gate."));
-
-        e = assertThrows(InvalidConfigurationException.class, () -> new FeatureGates("+UseKRaft,-KafkaNodePools"));
+        InvalidConfigurationException e = assertThrows(InvalidConfigurationException.class, () -> new FeatureGates("+UseKRaft,-KafkaNodePools"));
         assertThat(e.getMessage(), containsString("The UseKRaft feature gate can be enabled only together with the KafkaNodePools feature gate."));
     }
 
     @ParallelTest
     public void testKraftAndKafkaNodePoolsFulfilled() {
         assertThat(new FeatureGates("+UseKRaft,+KafkaNodePools").useKRaftEnabled(), is(true));
+        assertThat(new FeatureGates("+UseKRaft").useKRaftEnabled(), is(true));
     }
 }