+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.39.0
 
-
+* The `KafkaNodePools` feature gate moves to beta stage and is enabled by default.
+  If needed, `KafkaNodePools` can be disabled in the feature gates configuration in the Cluster Operator.
 
 ### Changes, deprecations and removals